+++ Familiar/Domain/Protocols/ScreenEdge.swift
@@ -1,0 +1,6 @@
+public enum ScreenEdge: Sendable {
+    case left
+    case right
+    case top
+    case bottom
+}