+++ Familiar/Domain/Protocols/SpriteProviding.swift
@@ -1,0 +1,7 @@
+public protocol SpriteProviding: AnyObject {
+    var frameCount: Int { get }
+    var frameWidth: Int { get }
+    var frameHeight: Int { get }
+    var isFlipped: Bool { get }
+    func flipAllFrames()
+}